requirements.txt
python==3.13.2
gzip
numpy
pillow
shutil
torch
torchvision